Refactor kopia wrapper module
