+++ pkg/kopia/repository/connect.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package kopia
+package repository
 
 import (
 	"context"
@@ -26,6 +26,7 @@ import (
 	"github.com/kopia/kopia/repo/content"
 	"github.com/pkg/errors"
 
+	"github.com/kanisterio/kanister/pkg/kopia"
 	"github.com/kanisterio/kanister/pkg/log"
 	"github.com/kanisterio/kanister/pkg/poll"
 )
@@ -47,7 +48,7 @@ func ConnectToAPIServer(
 	metadataCacheMB int,
 ) error {
 	// Extra fingerprint from the TLS Certificate secret
-	fingerprint, err := ExtractFingerprintFromCertificate(tlsCert)
+	fingerprint, err := kopia.ExtractFingerprintFromCertificate(tlsCert)
 	if err != nil {
 		return errors.Wrap(err, "Failed to extract fingerprint from Kopia API Server Certificate Secret")
 	}
@@ -61,7 +62,7 @@ func ConnectToAPIServer(
 
 	opts := &repo.ConnectOptions{
 		CachingOptions: content.CachingOptions{
-			CacheDirectory:            defaultCacheDirectory,
+			CacheDirectory:            kopia.DefaultCacheDirectory,
 			MaxCacheSizeBytes:         int64(contentCacheMB << 20),
 			MaxMetadataCacheSizeBytes: int64(metadataCacheMB << 20),
 			MaxListCacheDuration:      content.DurationSeconds(defaultConnectMaxListCacheDuration.Seconds()),
@@ -79,7 +80,7 @@ func ConnectToAPIServer(
 		Max:    15 * time.Second,
 	}, func(c context.Context) (bool, error) {
 		// TODO(@pavan): Modify this to use custom config file path, if required
-		err := repo.ConnectAPIServer(ctx, defaultConfigFilePath, serverInfo, userPassphrase, opts)
+		err := repo.ConnectAPIServer(ctx, kopia.DefaultConfigFilePath, serverInfo, userPassphrase, opts)
 		switch {
 		case isGetRepoParametersError(err):
 			log.Debug().WithError(err).Print("Connecting to the Kopia API Server")
