+++ pkg/kopia/utils.go
@@ -36,16 +36,11 @@ import (
 	"k8s.io/client-go/kubernetes"
 
 	"github.com/kanisterio/kanister/pkg/field"
+	snapshot2 "github.com/kanisterio/kanister/pkg/kopia/snapshot"
 	"github.com/kanisterio/kanister/pkg/log"
 )
 
 const (
-	// defaultConfigFilePath is the file which contains kopia repo config
-	defaultConfigFilePath = "/tmp/kopia-repository.config"
-
-	// defaultCacheDirectory is the directory where kopia content cache is created
-	defaultCacheDirectory = "/tmp/kopia-cache"
-
 	// defaultDataStoreGeneralContentCacheSizeMB is the default content cache size for general command workloads
 	defaultDataStoreGeneralContentCacheSizeMB = 0
 
@@ -144,8 +139,8 @@ func ExtractFingerprintFromCertificate(cert string) (string, error) {
 	return fingerprint, nil
 }
 
-// getStreamingFileObjectIDFromSnapshot returns the kopia object ID of the fs.StreamingFile object from the repository
-func getStreamingFileObjectIDFromSnapshot(ctx context.Context, rep repo.Repository, path, backupID string) (object.ID, error) {
+// GetStreamingFileObjectIDFromSnapshot returns the kopia object ID of the fs.StreamingFile object from the repository
+func GetStreamingFileObjectIDFromSnapshot(ctx context.Context, rep repo.Repository, path, backupID string) (object.ID, error) {
 	// Example: if the path from the blueprint is `/mysql-backups/1/2/mysqldump.sql`, the given backupID
 	// belongs to the root entry `/mysql-backups/1/2` with `mysqldump.sql` as a nested entry.
 	// The goal here is to find the nested entry and extract the object ID
@@ -197,7 +192,7 @@ func GetDataStoreGeneralMetadataCacheSize(opt map[string]int) int {
 }
 
 // MarshalKopiaSnapshot encodes kopia SnapshotInfo struct into a string
-func MarshalKopiaSnapshot(snapInfo *SnapshotInfo) (string, error) {
+func MarshalKopiaSnapshot(snapInfo *snapshot2.SnapshotInfo) (string, error) {
 	if err := snapInfo.Validate(); err != nil {
 		return "", err
 	}
@@ -210,8 +205,8 @@ func MarshalKopiaSnapshot(snapInfo *SnapshotInfo) (string, error) {
 }
 
 // UnmarshalKopiaSnapshot decodes a kopia snapshot JSON string into SnapshotInfo struct
-func UnmarshalKopiaSnapshot(snapInfoJSON string) (SnapshotInfo, error) {
-	snap := SnapshotInfo{}
+func UnmarshalKopiaSnapshot(snapInfoJSON string) (snapshot2.SnapshotInfo, error) {
+	snap := snapshot2.SnapshotInfo{}
 	if err := json.Unmarshal([]byte(snapInfoJSON), &snap); err != nil {
 		return snap, errors.Wrap(err, "failed to unmarshal kopia snapshot information")
 	}
